Changed Extraction Zone win condition + logic
Extraction zone is now found on level start instead of being hard coded into the UI. This allows level builders to place it wherever, and the UI will find it on it's own.

Also changed win condition to automatically start the extraction timer if all enemies are dead (instead of automatically winning).