Rewrote weapon script
Rewrote the entire weapon script. Definitely still has problems (projectile "firing").
Models not attatched, and player has to select a slot before the shoot & reload functions can be used. This includes after they first pick up a weapon as their slot is still technically slot 0. The player can not currently select a slot that does not have a weapon in it due to an index out of bounds error that I think it honestly fine but should be looked in to.